Change: Rename Header module to SectionHeader
Use the same name as the component and add tests.
