+++ src/web/pages/reports/AuditReportDetailsContent.tsx
@@ -23,8 +23,8 @@ import Divider from 'web/components/layout/Divider';
 import Layout from 'web/components/layout/Layout';
 import Loading from 'web/components/loading/Loading';
 import Powerfilter from 'web/components/powerfilter/PowerFilter';
-import SectionHeader from 'web/components/section/Header';
 import Section from 'web/components/section/Section';
+import SectionHeader from 'web/components/section/SectionHeader';
 import Tab from 'web/components/tab/Tab';
 import TabLayout from 'web/components/tab/TabLayout';
 import TabList from 'web/components/tab/TabList';
